refactor: Update logo image source in README.md
The code changes in this commit update the logo image source in the README.md file. The previous source was using an HTML `<img>` tag, and it has been replaced with a Markdown image syntax `![Logo FastAPI-HTMX](app/static/img/logo/logo.jpg)`. This change improves consistency and readability in the project documentation.