feat: :sparkles: Upload Documents is implemented in the application.
Upload documents is implemented in the application which enable user to upload the documents to Minio Object server, Download the Documents, Shows the list of documents uploaded by User. There is a upload model created with relationship to user to keep track the nature of documents. Unique names are provided to the document getting uploaded. Various front end interactive capabilities are performed